chore(server): extract shared cursor pagination primitives to common-api
Move the record-agnostic cursor primitives (decodeCursor, encodeCursorData,
getPaginationInfo) out of the dynamic-schema graphql-query-runner cursors util
into a neutral common-api util so they can be reused outside the object-record
pipeline (e.g. upcoming hand-written metadata resolvers). The record-specific
encodeCursor/processRecord connection logic stays put. Pure move + re-point of
importers; no behavior change.
