+++ packages/twenty-server/src/engine/api/common/utils/get-page-info.util.ts
@@ -3,10 +3,8 @@ import { type ObjectRecord } from 'twenty-shared/types';
 import { type ObjectRecordOrderBy } from 'src/engine/api/graphql/workspace-query-builder/interfaces/object-record.interface';
 
 import { type CommonPageInfo } from 'src/engine/api/common/types/common-page-info.type';
-import {
-  encodeCursor,
-  getPaginationInfo,
-} from 'src/engine/api/graphql/graphql-query-runner/utils/cursors.util';
+import { getPaginationInfo } from 'src/engine/api/common/utils/cursor.util';
+import { encodeCursor } from 'src/engine/api/graphql/graphql-query-runner/utils/cursors.util';
 import { type FlatEntityMaps } from 'src/engine/metadata-modules/flat-entity/types/flat-entity-maps.type';
 import { type FlatFieldMetadata } from 'src/engine/metadata-modules/flat-field-metadata/types/flat-field-metadata.type';
 import { type FlatObjectMetadata } from 'src/engine/metadata-modules/flat-object-metadata/types/flat-object-metadata.type';
