+++ packages/twenty-server/src/engine/api/graphql/graphql-query-runner/utils/__tests__/cursors.util.spec.ts
@@ -5,7 +5,10 @@ import { type ObjectRecordOrderBy } from 'src/engine/api/graphql/workspace-query
 import { type FlatEntityMaps } from 'src/engine/metadata-modules/flat-entity/types/flat-entity-maps.type';
 import { type FlatFieldMetadata } from 'src/engine/metadata-modules/flat-field-metadata/types/flat-field-metadata.type';
 import { type FlatObjectMetadata } from 'src/engine/metadata-modules/flat-object-metadata/types/flat-object-metadata.type';
-import { decodeCursor, encodeCursor } from '../cursors.util';
+
+import { decodeCursor } from 'src/engine/api/common/utils/cursor.util';
+
+import { encodeCursor } from '../cursors.util';
 
 const buildMockField = (
   id: string,
